refactor(create): fold agent-files back into src as TypeScript
The hand-written .mjs + .d.mts pair existed for one consumer: EdenX's
generator imported it through the `./agent-files` export. That generator now
ships its own copy so the two release independently, which leaves this package
as the only user — so the module moves into src as TypeScript, the export, the
typesVersions mapping and the hand-maintained declaration file all go, and the
types come from the compiler.
