+++ packages/toolkit/create/src/agent-files.ts
@@ -1,47 +1,28 @@
 import fs from 'node:fs';
 import path from 'node:path';
 
-// Writes the AGENTS.md / CLAUDE.md pair that points AI coding agents at a
-// framework's version-matched bundled docs.
-//
-// Everything is parameterized (marker name, managed block, target directory)
-// because frameworks built on Modern.js — EdenX, PIA — ship the same capability
-// from their own scaffolders and should reuse this rather than reimplement the
-// idempotency rules.
-//
-// Shipped as plain ESM so downstream packages can import it directly without
-// depending on how this package is bundled.
+// Writes the AGENTS.md / CLAUDE.md pair that points AI coding agents at the
+// version-matched bundled docs. Internal to this package: downstream
+// scaffolders (EdenX's generator) ship their own copy so the two release
+// independently.
+
+type FileOutcome = 'created' | 'updated' | 'added' | 'linked' | 'unchanged';
 
 const CLAUDE_IMPORT = '@AGENTS.md';
 
-const markers = name => ({
+const markers = (name: string) => ({
   begin: `<!-- BEGIN:${name} -->`,
   end: `<!-- END:${name} -->`,
 });
 
-/**
- * Extracts the managed block from a template file, so a scaffolder and this
- * codemod can share one source of truth for the block's content.
- *
- * @param {string} templateFile path to an AGENTS.md containing the markers
- * @param {string} markerName e.g. `modernjs-agent-rules`
- * @returns {string} the managed block, markers included
- */
-export function readManagedBlock(templateFile, markerName) {
-  const { begin, end } = markers(markerName);
-  const tpl = fs.readFileSync(templateFile, 'utf-8');
-  const from = tpl.indexOf(begin);
-  const to = tpl.indexOf(end);
-  if (from === -1 || to === -1 || to < from) {
-    throw new Error(`${templateFile} is missing the ${markerName} markers`);
-  }
-  return tpl.slice(from, to + end.length);
-}
-
 // Create AGENTS.md, refresh the managed block in place if present, or prepend
 // it — the "read the docs first" rule is the highest-priority instruction, so
 // it leads the file and the user's own content stays below it.
-function applyAgentsMd(targetDir, block, markerName) {
+function applyAgentsMd(
+  targetDir: string,
+  block: string,
+  markerName: string,
+): FileOutcome {
   const { begin, end } = markers(markerName);
   const file = path.join(targetDir, 'AGENTS.md');
 
@@ -70,7 +51,7 @@ function applyAgentsMd(targetDir, block, markerName) {
 
 // Claude Code reads CLAUDE.md, not AGENTS.md, so the import is what makes both
 // tools share a single set of instructions.
-function applyClaudeMd(targetDir) {
+function applyClaudeMd(targetDir: string): FileOutcome {
   const file = path.join(targetDir, 'CLAUDE.md');
 
   if (!fs.existsSync(file)) {
@@ -90,17 +71,13 @@ function applyClaudeMd(targetDir) {
   return 'linked';
 }
 
-/**
- * Idempotently writes AGENTS.md and CLAUDE.md into `targetDir`.
- *
- * @param {object} options
- * @param {string} options.targetDir project root to write into
- * @param {string} options.block managed block, markers included
- * @param {string} options.markerName marker name used inside `block`
- * @returns {{agents: string, claude: string}} what happened to each file:
- *   `created` | `updated` | `added` | `linked` | `unchanged`
- */
-export function applyAgentFiles({ targetDir, block, markerName }) {
+/** Idempotently writes AGENTS.md and CLAUDE.md into `targetDir`. */
+export function applyAgentFiles(options: {
+  targetDir: string;
+  block: string;
+  markerName: string;
+}): { agents: FileOutcome; claude: FileOutcome } {
+  const { targetDir, block, markerName } = options;
   if (!fs.existsSync(targetDir)) {
     throw new Error(`target directory does not exist: ${targetDir}`);
   }
